feat: atualiza controlador e views para gerenciamento de refeições, incluindo melhorias na edição e exclusão

+++ Controllers/RefeicaoController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 using ProjetoNutri.Context;
 using ProjetoNutri.Models;
 
@@ -19,16 +20,21 @@ namespace ProjetoNutri.Controllers
 
         public IActionResult IndexRefeicao()
         {
+            var refeicoes = _context.Refeicoes
+                .Include(r => r.Refeicao_Alimentos) // Inclui os alimentos da refeição
+                .ThenInclude(ra => ra.Alimento)
+                .OrderByDescending(r => r.DataCriacao)  // Se houver uma relação com Alimento, inclui também
+                .ToList();
 
-            var refeicoes = _context.Refeicoes.ToList();
             var viewModel = new ProjetoDietaViewModel
             {
                 Refeicaos = refeicoes
             };
 
-    // Retorne o ViewModel para a view
             return View(viewModel);
-        }  
+        }
+        
+
 
         public IActionResult CriarRefeicao()
         {
@@ -56,7 +62,7 @@ namespace ProjetoNutri.Controllers
         }
 
         [HttpPost]
-        public IActionResult EditarCategoria_Refeicao(Refeicao refeicao)
+        public IActionResult EditarRefeicao(Refeicao refeicao)
         {
             if (ModelState.IsValid)
             {
@@ -80,12 +86,19 @@ namespace ProjetoNutri.Controllers
         }
 
         [HttpPost]
-        public IActionResult DeletarCategoria_Refeicao(Refeicao refeicao)
+    public IActionResult Deletar_Refeicao(int id)
+    {
+        var refeicao = _context.Refeicoes.Find(id);
+        if (refeicao == null)
         {
-            _context.Refeicoes.Remove(refeicao);
-            _context.SaveChanges();
-            return RedirectToAction("IndexRefeicao");
+            return NotFound();
         }
+
+        _context.Refeicoes.Remove(refeicao);
+        _context.SaveChanges();
+        
+        return RedirectToAction("IndexRefeicao");
+    }
 
     }
 }
